requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
flask==3.0.3
flask-talisman==1.1.0
idna==3.10
importlib-metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.2
pluggy==1.5.0
prometheus-client==0.21.1
prometheus-flask-exporter==0.23.1
pytest==8.3.3
pytest-mock==3.14.0
python-dotenv==1.0.1
python-json-logger==2.0.7
requests==2.32.3
requests-mock==1.12.1
tomli==2.0.2
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2
